+++ tests/arm/fuzz_arm.cpp
@@ -191,7 +191,7 @@ void FuzzJitArm(const size_t instruction_count, const size_t instructions_to_exe
 
         // Run interpreter
         write_records.clear();
-        interp.NumInstrsToExecute = instructions_to_execute_count;
+        interp.NumInstrsToExecute = static_cast<unsigned>(instructions_to_execute_count);
         InterpreterMainLoop(&interp);
         auto interp_write_records = write_records;
         {
@@ -201,7 +201,7 @@ void FuzzJitArm(const size_t instruction_count, const size_t instructions_to_exe
 
         // Run jit
         write_records.clear();
-        jit.Run(instructions_to_execute_count);
+        jit.Run(static_cast<unsigned>(instructions_to_execute_count));
         auto jit_write_records = write_records;
 
         // Compare
